+++ Source Packages/se/moma/pryl/util/PersonVärdeComparator.java
@@ -7,7 +7,7 @@ public class PersonVärdeComparator implements Comparator<Person> {
     
     @Override
     public int compare(Person p1, Person p2) {
-	    return p2.summaVärde() - p1.summaVärde();
+	    return (int) (p2.summaVärde() - p1.summaVärde());
     }
 }
